Fix build process
Instead of building the libDBoW2, libg2o and libORB_SLAM3 in <repo>/lib,
build it in the build directory as per standard in many cmake projects.
i.e the built libraries are in:

    ./build/Thirdpart/DBoW2/libDBoW2.so
    ./build/Thirdpart/g2o/libg2o.so
    ./build/libORB_SLAM3.so

This commit also removes the need for the build.sh script, meaning all
the user has to do is the standard cmake build procedure:

    mkdir -p build
    cd build
    cmake .. -DCMAKE_BUILD_TYPE=Release
    make

This is much preferred because this is the widely accepted convention in
building a cmake project.